image_generator/requirements.txt
crewai
python-dotenv
openai
requests
diffusers
torch
transformers
pyyaml
accelerate